+++ sunpy/timeseries/sources/eve.py
@@ -47,7 +47,7 @@ class EVESpWxTimeSeries(GenericTimeSeries):
     >>> import sunpy.data.sample
 
     >>> eve = sunpy.timeseries.TimeSeries(sunpy.data.sample.EVE_TIMESERIES, source='EVE')
-    >>> eve = sunpy.timeseries.TimeSeries("http://lasp.colorado.edu/eve/data_access/quicklook/quicklook_data/L0CS/LATEST_EVE_L0CS_DIODES_1m.txt", source='EVE')
+    >>> eve = sunpy.timeseries.TimeSeries("http://lasp.colorado.edu/eve/data_access/quicklook/quicklook_data/L0CS/LATEST_EVE_L0CS_DIODES_1m.txt", source='EVE')  # doctest: +REMOTE_DATA
     >>> eve.peek(subplots=True)    # doctest: +SKIP
 
     References